fix maximum ploop version calculation
= Problem

Before migrate a CT to a different node we should check that kernel and
userspace of that node supports all CT's ploop disks formats. This
check in current vzmigrate is outdated and doesn't supports external
disks, specifically it expects that root disk resides in CT bundle.

= Solution

Fix to support external disks.

review: @110752
bug: #PSBM-39290