BUGFIX Improve socketname generation method in compat mode #PSBM-60800
In VZ6 vzmdest was appending CTIDs to the socketname, therefore migration
of a UUID-named CT in VZ7 could easily reach maximum size for the unix
socket filename which is 108 symbols. In VZ7 socketname generation method
was changed in order to shorten the filename, but it broke compatibility
with VZ6. If revert the mothod to vz6-like, but, trim ID to the first dash
symbol we will easily fit into the limit and maintain compatibility with
VZ6 since there is no dashes in CTID.