bug #PSBM-65829 aka CT not visible for "prlctl list" after "vzmlocal -C"
we have a race between setting the clone ct name in config and
handling a registered event inside dispatcher. the latter sees the
name of the clone source and refuses to register another ct with
the same name.
let's revert to the old scheme when config from /etc/vz/conf replaces
the one inside the clone ct private. thus we update the config
before registration not after.