Allow online migration of stopped containers in Agent mode #PSBM-66166
In latest VA MN it was decided to mgirate in online mode by default for
the sake of usability and to eliminate accidental downtime. Vzmigrate in
its turn does not allow to migrate stopped CT with "--online" option,
lets allow it, and clean "OPT_ONLINE" flag for stopped containers when
migrating in Agent mode.